requirements.txt
asgiref==3.8.1
Django==5.1
django-cors-headers==4.4.0
django-filter==24.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
gunicorn==23.0.0
Markdown==3.6
packaging==24.1
psycopg2==2.9.9
PyJWT==2.9.0
sqlparse==0.5.1
tzdata==2024.1
whitenoise==6.7.0
